REFACTOR: add user to utils
